fix: detail view for canned response and refactor
