+++ desk/src/components/canned-response/CannedResponseInfo.vue
