+++ desk/src/components/canned-response/index.js
@@ -1,0 +1,2 @@
+export { default as AddNewCannedResponseDialog } from "./AddNewCannedResponseDialog.vue";
+export { default as CannedResponseInfo } from "./CannedResponseInfo.vue";
